Nested saving + refactor.
