+++ tests/material/dumps/compound_a.txt
@@ -1,0 +1,27 @@
+from lf import trans
+
+
+CONST = 1
+
+
+def k(o):
+    return o * 7
+
+
+@trans
+def x(y):
+    return CONST + 1
+
+
+@trans
+def y(y):
+    return CONST + k(y)
+
+
+_lf_main = (
+    (
+        x
+        >> y
+    )
+    >> x
+)
